requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.0
colorama==0.4.6
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gotrue==2.12.4
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==10.4.0
postgrest==0.16.11
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.7
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==12.0
Werkzeug==3.1.3
WTForms==3.2.1
